feat: vistiro info pop up
